+++ Assets/SCOB/DataRepo/Data.cs
@@ -4,26 +4,26 @@ using UnityEngine;
 [CreateAssetMenu(menuName = "Objects/DataRepo")]
 public class Data : ScriptableObject
 {
-    public int BottleColorRandomIndex = 0;
-    public int ExtraBottleAmountRandomIndex = 0;
-    public int ColorPickerRandomIndex = 0;
+    public long BottleColorRandomIndex = 0;
+    public long ExtraBottleAmountRandomIndex = 0;
+    public long ColorPickerRandomIndex = 0;
 
     [SerializeField] public List<Bottle> CreatedBottles = new List<Bottle>();
 
 
-    public int GetBottleColorRandomIndex()
+    public long GetBottleColorRandomIndex()
     {
         BottleColorRandomIndex += 50;
         return BottleColorRandomIndex - 50;
     }
 
-    public int GetAmountOfExtraBottleIndex()
+    public long GetAmountOfExtraBottleIndex()
     {
         ExtraBottleAmountRandomIndex += 50;
         return ExtraBottleAmountRandomIndex - 50;
     }
 
-    public int GetColorPickerRandomIndex()
+    public long GetColorPickerRandomIndex()
     {
         ColorPickerRandomIndex +=50;
         return ColorPickerRandomIndex - 50;
